convert simplehttpserver to modular
