+++ src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module simplehttpserver {
+
+    requires jdk.httpserver;
+    requires java.net.http; // test requirement
+    exports com.kttdevelopment.simplehttpserver.handler;
+    exports com.kttdevelopment.simplehttpserver.var;
+    exports com.kttdevelopment.simplehttpserver;
+
+}
